handle err when translating path
this can happen when a file that looks like a whiteout on the underlying
driver is present in the volume; apparently it's listed as a dir entry
but then statting it fails, probably because the filesystem is trying
deparately to make the file appear removed. unsure why it was listed in
the first place, though.

(context: whiteouts on overlay are 0,0 char device nodes.)