volume: add read-only registry endpoints
this commit adds three new endpoints to baggageclaim's api

	/v2				--> status check
	/v2/<name>/manifests/<ref>	--> img manifest
	/v2/<name>/blobs/<digest>	--> layers, etc (blobs)

whose responsability is to serve the contents of a volume as a container
images, following the distribution spec (1).

the way that this is intended to work is as follows:

	  i. a volume is created, and populated with a container image
	     (e.g., from a `get` step from `registry-image`: image.tar)

	 ii. someone wanting to consume that volume as an image points
	     to baggageclaim

		docker pull baggageclaim:7788/handle:latest

	iii. the handlers then deal with that `image.tar`, first serving
	     the manifest, and then the corresponding blobs.

given that *any* volume that has an `image.tar` would be able to serve
its contents as container images, this means that we could even have
containers whose whole responsability is to serve images that it makes
servable.

(1): https://github.com/opencontainers/distribution-spec

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>