make concourse's img format serve'able
by bringing a new implementation of the `Image` interface, we can get
concourse's own image format able to be served as a container image by
performing the necessary conversions.

the approach here is to create the config, manifest, and rootfs
representations on disk, making that all loadable only if necessary.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>