fix depth >2 overlay cow volumes
when creating a copy of a copy, instead of nesting, copy the parent's
upper dir into the child's upper dir and then set the child's parent to
the grandparent.

this keeps the depth from growing unbounded, which is important
considering the kernel has a hardcoded depth limit of 2. the trade-off
is that we'll copy the upper directory contents in this situation, but
hopefully the upper directory will usually be small enough for this to
not matter much.

(in any case, there doesn't seem to be a better choice. sadly.)

concourse/concourse#5799

Signed-off-by: Alex Suraci <asuraci@vmware.com>
Co-authored-by: Vikram Yadav <vyadav@vmware.com>