fix nesting volumes even more deeply
the previous fix actually only made it work slightly more deeply; this
should fix it all the way. the problem was that the grandparent was
found using its logical parent, rather than the 'actual' parent, which
must be found by recursively walking through the hierarchy all the way
to the bottom.

* test up to depth 10
* assert that file updates and deletions are preserved at each depth,
  just to be sure

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>