add prs pipeline
I'm working on creating a separate PR pipeline for master vs. release
branches on concourse/concourse. As part of that, I'm moving
baggageclaim and docs PRs to their own pipelines rather than putting
them in the "main" PRs pipeline.

This way, we can create copies of the same concourse PR pipeline for
different branches without running baggageclaim/docs PRs multiple times.
This is also necessary for when we use instanced pipelines for PRs

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>