fix 'invalid argument' flake
the overlay tests would only pass if the image was fetched on another
worker and then streamed to the worker that runs the tests, which would
result in / being ext4 (or whatever the host filesystem is).

if the image was already present on the worker, it'd be a copy-on-right
volume, which would result in / (and thus /tmp) being overlay, so the
overlay tests failing with 'invalid argument' and a dmesg error saying
the filesystem is not supported.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>