convert baggageclaim to use cobra
As part of the epic for having concourse web be deployed using a config
file, we converted all the commands to be using cobra. Baggageclaim
command will be converted to become part of a config file and there will
also be a separate cobra command that can be used if you want to start
baggageclaim up independently from the worker.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>