minor changes and fixes
- add a supplied prefix to flags because baggageclaim will be used by
the worker and quickstart command in concourse repo. The quickstart
command prefixes all the worker flags with "worker-"
- Add "inline" to the yaml tags for when we don't want to use the
field name within the yaml config structure
- Remove supporting fetching of configuration fields through environment
variable when running baggageclaim independently (not part of
concourse). This was never actually a supported feature, I just thought
it was but realized that it was not.

[#6434]

Signed-off-by: Clara Fu <fclara@vmware.com>